requirements.txt
# Core ML/DL libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# EEG processing
mne>=1.5.0
mne-bids>=0.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Evaluation metrics
nltk>=3.8.0
rouge-score>=0.1.2
sacrebleu>=2.3.0

# Configuration and logging
pyyaml>=6.0
wandb>=0.15.0
tensorboard>=2.13.0

# Utilities
tqdm>=4.65.0
argparse>=1.4.0
pathlib>=1.0.1
joblib>=1.3.0

# Optional: For advanced features
# optuna>=3.2.0  # Hyperparameter optimization
# ray[tune]>=2.6.0  # Distributed training
# deepspeed>=0.9.0  # Memory optimization 